requirements.txt
# Web框架
fastapi==0.115.8
python-multipart==0.0.20
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.0

# AI/LLM
dashscope==1.20.0

# 对象存储
minio==7.2.0

# 数据库
SQLAlchemy==2.0.38
alembic==1.14.1

# HTTP请求
requests==2.31.0

# PDF生成
reportlab==4.0.7
pillow==10.1.0

uvicorn==0.35.0
python-jose[cryptography]==3.5.0
eth-account==0.13.7
